feat: Opening a file and printing it's content to stdout
Taking the arguments using argv to fetch for the file in the position
argv[1]. If the file dosen't exist an error message is printed to stderr.

Looping throw the file and printing the it's content line by line.
Also keeping track of the line number to use later for error handling.

Missing features for basic code are:
- Styling the line to get the opcode and the argument if necessary.
- Handle errors if the command wasn't found or is missing an argument.
- Create a function that check if the given opcode is equal to builtin
function. ( builtin functions will be stored in an array )